+++ src/interp.cc
@@ -13,19 +13,19 @@ void Interpreter::Run(Heap &heap, Instr *code) {
     case Instr::Code::NOP: {
     } break;
     case Instr::Code::INCR_CELL: {
-      heap.IncrementCell(code->Operand1());
+      heap.IncrementCell((uint8_t) code->Operand1());
     } break;
     case Instr::Code::DECR_CELL: {
-      heap.DecrementCell(code->Operand1());
+      heap.DecrementCell((uint8_t) code->Operand1());
     } break;
     case Instr::Code::SET_CELL: {
       heap.SetCell((uint8_t) code->Operand1());
     } break;
     case Instr::Code::INCR_PTR: {
-      heap.IncrementDataPointer(code->Operand1());
+      heap.IncrementDataPointer((int64_t) code->Operand1());
     } break;
     case Instr::Code::DECR_PTR: {
-      heap.DecrementDataPointer(code->Operand1());
+      heap.DecrementDataPointer((int64_t) code->Operand1());
     } break;
     case Instr::Code::READ: {
       const uint8_t input = (uint8_t) std::getchar();
@@ -48,13 +48,13 @@ void Interpreter::Run(Heap &heap, Instr *code) {
     case Instr::Code::FIND_CELL_HIGH: {
       const uint8_t val = (uint8_t) code->Operand1();
       while (heap.GetCell() != val) {
-        heap.IncrementDataPointer(code->Operand2());
+        heap.IncrementDataPointer((int64_t) code->Operand2());
       }
     } break;
     case Instr::Code::FIND_CELL_LOW: {
       const uint8_t val = (uint8_t) code->Operand1();
       while (heap.GetCell() != val) {
-        heap.DecrementDataPointer(code->Operand2());
+        heap.DecrementDataPointer((int64_t) code->Operand2());
       }
     } break;
     }
